Fit the Bradley-Terry model using the EM or MM algorithm

Fit the Bradley-Terry model using the EM or MM algorithm

BT_EM(W, a, b, maxit = 5000L, epsilon = 0.001)

Arguments

W
a K*K square matrix of class "dgCMatrix"

a
the shape parameter of the gamma prior

b
the rate parameter of the gamma prior

maxit
the maximum number of iterations

epsilon
controls the convergence criteria

Value

A list containing a K*1 matrix with the pi estimate, the N matrix, the number of iterations, and whether the algorithm converged.