Calculates Bradley-Terry probabilities

Calculates the Bradley-Terry probabilities of each item in a fully-connected component of \(G_W\) winning against every other item in that component (see Details).

btprob(object, subset = NULL, as_df = FALSE)

Arguments

object
An object of class "btfit", typically the result ob of ob <- btfit(..). See btfit.

subset
A condition for selecting a subset of the components. This can either be a character vector of names of the components (i.e. a subset of names(object$pi)), a single predicate function (that takes a vector of object$pi as its argument), or a logical vector of the same length as the number of components, (i.e. length(object$pi)).

as_df
Logical scalar, determining class of output. If TRUE, the function returns a data frame. If FALSE (the default), the function returns a matrix (or list of matrices).

Value

If as_df = FALSE, returns a matrix where the \(i,j\)-th element is the Bradley-Terry probability \(p_{ij}\), or, if \(G_W\) is not fully-connected and btfit has been run with a = 1, a list of such matrices for each fully-connected component of \(G_W\). If as_df = TRUE, returns a five-column data frame, where the first column is the component that the two items are in, the second column is item1, the third column is item2, the fourth column is the Bradley-Terry probability that item 1 beats item 2 and the fifth column is the Bradley-Terry probability that item 2 beats item 1. If the original btdata$wins matrix has named dimnames, these will be the colnames for columns one and two. See Details.

Details

Consider a set of \(K\) items. Let the items be nodes in a graph and let there be a directed edge \((i, j)\) when \(i\) has won against \(j\) at least once. We call this the comparison graph of the data, and denote it by \(G_W\). Assuming that \(G_W\) is fully connected, the Bradley-Terry model states that the probability that item \(i\) beats item \(j\) is $$p_{ij} = \frac{\pi_i}{\pi_i + \pi_j},$$ where \(\pi_i\) and \(\pi_j\) are positive-valued parameters representing the skills of items \(i\) and \(j\), for \(1 \le i, j, \le K\). The function btfit can be used to find the strength parameter \(\pi\). It produces a "btfit" object that can then be passed to btprob to obtain the Bradley-Terry probabilities \(p_{ij}\).

If \(G_W\) is not fully connected, then a penalised strength parameter can be obtained using the method of Caron and Doucet (2012) (see btfit, with a > 1), which allows for a Bradley-Terry probability of any of the K items beating any of the others. Alternatively, the MLE can be found for each fully connected component of \(G_W\) (see btfit, with a = 1), and the probability of each item in each component beating any other item in that component can be found.

References

Bradley, R. A. and Terry, M. E. (1952). Rank analysis of incomplete block designs: 1. The method of paired comparisons. Biometrika, 39(3/4), 324-345.

Caron, F. and Doucet, A. (2012). Efficient Bayesian Inference for Generalized Bradley-Terry Models. Journal of Computational and Graphical Statistics, 21(1), 174-196.

See also

btfit, btdata

Examples

W_connected <- Matrix::rsparsematrix(10, 10 , 0.5, rand.x = function(n) rbinom(n, 10, 0.5))
i <- c(3,1,5,4,2,5,5,7,8,5,6,8,7)
j <- c(1,2,2,3,4,4,6,6,6,7,7,7,8)
dimnames = list(letters[1:8], letters[1:8])
W_not_connected <-  Matrix::sparseMatrix(i, j, x = 1:13, dims = c(8,8), dimnames = dimnames)
W_connected_data <- btdata(W_connected)
W_not_connected_data <- btdata(W_not_connected)
fit1 <- btfit(W_connected_data, 1)
fit2 <- btfit(W_not_connected_data, 1)
fit3 <- btfit(W_not_connected_data, 3)
btprob(fit1)

#> 10 x 10 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 10 column names ‘7’, ‘10’, ‘9’ ... ]]

#>                                                                          
#> 7  .          0.7635455 0.7692865 0.8796861 0.8885760 0.9132446 0.9137278
#> 10 0.23645446 .         0.5080167 0.6936510 0.7117832 0.7652522 0.7663489
#> 9  0.23071349 0.4919833 .         0.6867942 0.7051597 0.7594422 0.7605576
#> 2  0.12031388 0.3063490 0.3132058 .         0.5216905 0.5901171 0.5915954
#> 8  0.11142399 0.2882168 0.2948403 0.4783095 .         0.5689659 0.5704649
#> 3  0.08675544 0.2347478 0.2405578 0.4098829 0.4310341 .         0.5015287
#> 1  0.08627218 0.2336511 0.2394424 0.4084046 0.4295351 0.4984713 .        
#> 4  0.07992455 0.2190595 0.2245952 0.3884312 0.4092432 0.4776494 0.4791753
#> 6  0.05324818 0.1537020 0.1579200 0.2913963 0.3096415 0.3718794 0.3733089
#> 5  0.03554335 0.1063485 0.1094350 0.2122609 0.2271395 0.2795084 0.2807416
#>                                 
#> 7  0.9200754 0.9467518 0.9644567
#> 10 0.7809405 0.8462980 0.8936515
#> 9  0.7754048 0.8420800 0.8905650
#> 2  0.6115688 0.7086037 0.7877391
#> 8  0.5907568 0.6903585 0.7728605
#> 3  0.5223506 0.6281206 0.7204916
#> 1  0.5208247 0.6266911 0.7192584
#> 4  .         0.6069956 0.7021254
#> 6  0.3930044 .         0.6041384
#> 5  0.2978746 0.3958616 .        

btprob(fit2)

#> $`2`
#> 3 x 3 sparse Matrix of class "dgCMatrix"
#>           h         g         f
#> h .         0.5985400 0.6704501
#> g 0.4014600 .         0.5770893
#> f 0.3295499 0.4229107 .        
#> 
#> $`3`
#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>            a          b         d         c
#> a .          0.52920629 0.8287550 0.9402557
#> b 0.47079371 .          0.8115130 0.9333372
#> d 0.17124498 0.18848704 .         0.7648125
#> c 0.05974435 0.06666283 0.2351875 .        
#> 

btprob(fit2, as_df = TRUE)

#> # A tibble: 9 x 5
#>   component item1 item2 prob1wins  prob2wins
#>       <chr> <chr> <chr>     <dbl>      <dbl>
#> 1         2     h     g 0.5985400 0.40146002
#> 2         2     h     f 0.6704501 0.32954991
#> 3         2     g     f 0.5770893 0.42291068
#> 4         3     a     b 0.5292063 0.47079371
#> 5         3     a     d 0.8287550 0.17124498
#> 6         3     b     d 0.8115130 0.18848704
#> 7         3     a     c 0.9402557 0.05974435
#> 8         3     b     c 0.9333372 0.06666283
#> 9         3     d     c 0.7648125 0.23518746

btprob(fit3)

#> 8 x 8 sparse Matrix of class "dgCMatrix"
#>           e         a         b         h         d         g         c
#> e .         0.7303944 0.7478675 0.7794295 0.8452451 0.8474738 0.8696628
#> a 0.2696056 .         0.5226461 0.5660419 0.6684448 0.6722324 0.7112281
#> b 0.2521325 0.4773539 .         0.5436570 0.6480578 0.6519570 0.6922607
#> h 0.2205705 0.4339581 0.4563430 .         0.6071717 0.6112522 0.6537663
#> d 0.1547549 0.3315552 0.3519422 0.3928283 .         0.5042848 0.5498827
#> g 0.1525262 0.3277676 0.3480430 0.3887478 0.4957152 .         0.5456369
#> c 0.1303372 0.2887719 0.3077393 0.3462337 0.4501173 0.4543631 .        
#> f 0.1229178 0.2751875 0.2936309 0.3312049 0.4335698 0.4377838 0.4832302
#>           f
#> e 0.8770822
#> a 0.7248125
#> b 0.7063691
#> h 0.6687951
#> d 0.5664302
#> g 0.5622162
#> c 0.5167698
#> f .