Fitted Method for "btfit"

fitted.btfit returns the fitted values from a fitted btfit model object.

# S3 method for btfit
fitted(object, subset = NULL, as_df = FALSE, ...)

Arguments

object
An object of class "btfit", typically the result ob of ob <- btfit(..). See btfit.

subset
A condition for selecting a subset of the components. This can either be a character vector of names of the components (i.e. a subset of names(object$pi)), a single predicate function (that takes a vector of object$pi as its argument), or a logical vector of the same length as the number of components, (i.e. length(object$pi)).

as_df
Logical scalar, determining class of output. If TRUE, the function returns a data frame. If FALSE (the default), the function returns a matrix (or list of matrices).

...
Other arguments

Value

If as_df = FALSE and the model has been fit on the full dataset, returns a matrix where the \(i,j\)-th element is the Bradley-Terry expected value \(m_{ij}\). Otherwise, a list of such matrices is returned, one for each fully-connected component. If as_df = TRUE, returns a five-column data frame, where the first column is the component that the two items are in, the second column is item1, the third column is item2, the fourth column, fit1, is the expected number of times that item 1 beats item 2 and the fifth column, fit2, is the expected number of times that item 2 beats item 1. If btdata$wins has named dimnames, these will be the colnames for columns one and two. Otherwise these colnames will be item1 and item2. See Details.

Details

Consider a set of \(K\) items. Let the items be nodes in a graph and let there be a directed edge \((i, j)\) when \(i\) has won against \(j\) at least once. We call this the comparison graph of the data, and denote it by \(G_W\). Assuming that \(G_W\) is fully connected, the Bradley-Terry model states that the probability that item \(i\) beats item \(j\) is $$p_{ij} = \frac{\pi_i}{\pi_i + \pi_j},$$ where \(\pi_i\) and \(\pi_j\) are positive-valued parameters representing the skills of items \(i\) and \(j\), for \(1 \le i, j, \le K\).

The expected, or fitted, values under the Bradley-Terry model are therefore: $$m_{ij} = n_{ij}p_{ij},$$

where \(n_ij\) is the number of times item \(i\) plays against item \(j\).

The function btfit is used to fit the Bradley-Terry model. It produces a "btfit" object that can then be passed to fitted.btfit to obtain the fitted values \(m_{ij}\). Note that the Bradley-Terry probabilities \(p_{ij}\) can be calculated using btprob.

If \(G_W\) is not fully connected, then a penalised strength parameter can be obtained using the method of Caron and Doucet (2012) (see btfit, with a > 1), which allows for a Bradley-Terry probability of any of the K items beating any of the others. Alternatively, the MLE can be found for each fully connected component of \(G_W\) (see btfit, with a = 1), and the probability of each item in each component beating any other item in that component can be found.

References

Bradley, R. A. and Terry, M. E. (1952). Rank analysis of incomplete block designs: 1. The method of paired comparisons. Biometrika, 39(3/4), 324-345.

Caron, F. and Doucet, A. (2012). Efficient Bayesian Inference for Generalized Bradley-Terry Models. Journal of Computational and Graphical Statistics, 21(1), 174-196.

See also

btfit, btprob

Examples

W_connected <- Matrix::rsparsematrix(10, 10 , 0.5, rand.x = function(n) rbinom(n, 10, 0.5))
i <- c(3,1,5,4,2,5,5,7,8,5,6,8,7)
j <- c(1,2,2,3,4,4,6,6,6,7,7,7,8)
dimnames = list(letters[1:8], letters[1:8])
W_not_connected <-  Matrix::sparseMatrix(i, j, x = 1:13, dims = c(8,8), dimnames = dimnames)
W_connected_data <- btdata(W_connected)
W_not_connected_data <- btdata(W_not_connected)
fit1 <- btfit(W_connected_data, 1)
fit2 <- btfit(W_not_connected_data, 1)
fit3 <- btfit(W_not_connected_data, 3)
fitted(fit1)

#> 10 x 10 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 10 column names ‘3’, ‘2’, ‘9’ ... ]]

#>                                                                             
#> 3  .         6.3358771 .        6.806368 4.676736 .        7.644278 5.960773
#> 2  3.6641229 6.0000000 6.248298 3.210696 4.700319 6.752192 7.653044 4.610171
#> 9  .         4.7517015 5.000000 .        6.085100 7.963263 .        .       
#> 10 2.1936316 1.7893040 .        4.000000 3.727527 .        5.160350 7.138439
#> 1  1.3232638 2.2996810 3.914900 3.272473 .        3.025540 7.991138 3.712469
#> 6  .         3.2478081 5.036737 .        2.974460 3.000000 1.831987 2.458880
#> 5  1.3557215 2.3469558 .        2.839650 5.008862 1.168013 4.000000 2.521374
#> 4  1.0392270 1.3898295 .        3.861561 2.287531 1.541120 2.478626 .       
#> 8  .         .         .        .        2.541615 .        2.381511 1.441707
#> 7  0.3997987 0.8401206 1.813316 .        .        3.559796 1.857483 5.154889
#>                     
#> 3  .        2.600201
#> 2  .        3.159879
#> 9  .        5.186684
#> 10 .        .       
#> 1  4.458385 .       
#> 6  .        6.440204
#> 5  2.618489 2.142517
#> 4  1.558293 5.845111
#> 8  6.000000 5.631691
#> 7  5.368309 .       

fitted(fit2)

#> $`2`
#> 3 x 3 sparse Matrix of class "dgCMatrix"
#>           h         g         f
#> h  .        14.963499  6.034051
#> g 10.036501  .        10.964697
#> f  2.965949  8.035303  .       
#> 
#> $`3`
#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>            a         b         d         c
#> a .          1.0584126 .         0.9402557
#> b 0.94158742 .         4.0575648 .        
#> d .          0.9424352 .         3.0592502
#> c 0.05974435 .         0.9407498 .        
#> 

fitted(fit2, as_df = TRUE)

#> # A tibble: 7 x 5
#>   component item1 item2       fit1        fit2
#>       <chr> <chr> <chr>      <dbl>       <dbl>
#> 1         2     h     g 14.9634994 10.03650056
#> 2         2     h     f  6.0340508  2.96594922
#> 3         2     g     f 10.9646972  8.03530284
#> 4         3     a     b  1.0584126  0.94158742
#> 5         3     b     d  4.0575648  0.94243521
#> 6         3     a     c  0.9402557  0.05974435
#> 7         3     d     c  3.0592502  0.94074985

fitted(fit3)

#> 8 x 8 sparse Matrix of class "dgCMatrix"
#>           e         a        b        h        d         g         c         f
#> e .         .         2.243602 .        5.071470  8.474738 .          6.139576
#> a .         .         1.045292 .        .         .        0.7112281  .       
#> b 0.7563976 0.9547078 .        .        3.240289  .        .          .       
#> h .         .         .        .        .        15.281305 .          6.019156
#> d 0.9285295 .         1.759711 .        .         .        2.1995308  .       
#> g 1.5252618 .         .        9.718695 .         .        .         10.682109
#> c .         0.2887719 .        .        1.800469  .        .          .       
#> f 0.8604244 .         .        2.980844 .         8.317891 .          .