Converts a data frame of paired results into a square matrix.

pairs_to_matrix is a data manipulation function that converts a data frame of paired results (either ordered or unordered) into a square matrix. This is a helper function for btdata.

pairs_to_matrix(df)

Arguments

df
A data frame (either a data.frame or an object that inherits that class, such as a tbl_df or a data.table). It must have 3 or 4 columns:

Ordered

If a 3-column data-frame, the first column contains the name of the winning player, the second column contains the name of the losing player and the third columns contains the number of times that the winner has beaten the loser. Multiple entries for the same pair of teams are handled correctly.

Unordered

If a 4-column data-frame, the first column contains the name of Player 1, the second column contains the name of Player 2, the third column contains the number of times that Player 1 has beaten Player 2 and the fourth column contains the number of times Player 2 has beaten Player 1. Multiple entries for the same pair of teams are handled correctly.

In both cases, if there is a repeated pairing in df, pairs_to_matrix aggregates the number of wins, as one would expect.

Value

A \(K\) by \(K\) sparse Matrix, where \(K\) is the number of players. The \(i,j\)-th element of this matrix is the number of times player \(i\) beats player \(j\).

See also

btdata.

Examples

winner <- c("A", "A", "A", "A", "A", "B", "B", "B", "C", "C", "E", "F", "F")
loser <- c("B", "B", "C", "D", "F", "D", "A", "F", "F", "G", "G", "A", "C")
df <- data.frame(winner = winner, loser = loser, num_wins = 1:13)
pairs_to_matrix(df)

#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>       loser
#> winner  A B  C D E F  G
#>      A  . 3  3 4 . 5  .
#>      B  7 .  . 6 . 8  .
#>      C  . .  . . . 9 10
#>      D  . .  . . . .  .
#>      E  . .  . . . . 11
#>      F 12 . 13 . . .  .
#>      G  . .  . . . .  .

df2 <- data.frame(team1 = c("A", "A", "B", "C"),
                  team2 = c("B", "C", "C", "D"),
                  wins1 = 1:4, wins2 = 5:8)
pairs_to_matrix(df2)

#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>      team2
#> team1 A B C D
#>     A . 1 2 .
#>     B 5 . 3 .
#>     C 6 7 . 4
#>     D . . 8 .