Simulate data from a specified Bradley-Terry model

Simulate data from a specified Bradley-Terry model

simulate_BT(pi, N, nsim = 1, seed = NULL, result_class = c("sparseMatrix",
  "btdata"))

# S3 method for btfit
simulate(object, nsim = 1, seed = NULL,
  result_class = c("sparseMatrix", "btdata"), ...)

Arguments

pi
a numeric vector, with all values finite and positive.

N
a symmetric, numeric matrix with dimensions the same as length(pi). The elements should be numeric representations of non-negative integers.

nsim
a scalar integer, the number of datasets to be generated.

seed
an object specifying if and how the random number generator should be initialized (‘seeded’). For details see simulate.

result_class
a character vector specifying whether the generated datasets should be of class "sparseMatrix" or of class "btdata". If not specified, the first match among those alternatives is used.

object
An object of class "btfit", typically the result of ob of ob <- btfit(..). This object must only have one component, i.e. length(object$pi) == 1.

...
Other arguments

Value

a list of length nsim of simulated datasets. If result_class = "sparseMatrix", the datasets are sparse matrices with the same dimensions as N. If result_class = "btdata" then the datasets are "btdata" objects. See btdata

See also

btfit, btdata

Examples

citations_btdata <- btdata(BradleyTerryScalable::citations)
fit1 <- btfit(citations_btdata, 1)
simulate(fit1, nsim = 2)

#> $sim_1
#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>               citing
#> cited          JRSS-B Biometrika JASA Comm Statist
#>   JRSS-B            .        302  301          289
#>   Biometrika      203          .  500          725
#>   JASA            166        318    .          809
#>   Comm Statist      4         38   72            .
#> 
#> $sim_2
#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>               citing
#> cited          JRSS-B Biometrika JASA Comm Statist
#>   JRSS-B            .        285  319          287
#>   Biometrika      220          .  516          724
#>   JASA            148        302    .          809
#>   Comm Statist      6         39   72            .
#> 
#> attr(,"seed")
#>   [1]         403         624  -207409486 -1849856373 -1042705328  -602680303
#>   [7]   297184926   616946311  1494210844  -120371347  1128119114 -1237599165
#>  [13] -1297602392 -1017047415  2037186742  -972501057  -681963788   349461861
#>  [19]  -734926110  1606979835  1945360384   339474433   933144014  1044659191
#>  [25] -1884909620   226132573 -2045371014 -1666265933   621937240 -1679665031
#>  [31] -1422393882  -241107153 -1441743964  -997091755  1670165266 -1881845909
#>  [37]  1380190128  1552816113  1827760382  -391906969 -1738542468  -694060723
#>  [43] -1950086230  -331417309  1513572360  1425348201 -1865464042 -1012004193
#>  [49] -1819831212 -1704064187  1274447682  -527783461 -2088426656   935932897
#>  [55]   396490798   511447767  -920190676   276245565  1760200154  1920156051
#>  [61] -1036846664   322457689 -1870275514  1671885327   956930308 -1032274891
#>  [67] -1737332878  1246732363   758262544  -395555887  -362589346   270764103
#>  [73]  1118223324 -1830118611   720145418  -292379133   601888616   856360521
#>  [79]  1970191734  1724036479  -492713548  2088806693  -456961118  1867199163
#>  [85]   595992256  1622564801   346585742 -1817093705 -1605754228  1101710877
#>  [91]   132980282 -2119948685  1289330968   948713529 -1621300570  1073239279
#>  [97]   723199588   202691093 -1922282542   191127851 -1719927184   849768369
#> [103]  1923378622 -1795389657 -1472449220   -64574195 -1899021206   510575331
#> [109] -1048930616 -1080367575   853763030 -1342255009 -1342132460 -1467730171
#> [115]  -542093310  1682058139  -791754208  2147187617 -1119212306 -2053369705
#> [121]   -77073428 -1920118787  -684333414  -109479085  1826486392  1809190937
#> [127]  1230317830   894248911 -1046649916  1826478069  1234329650 -1186229749
#> [133]  -906395184  -365657199 -1169377250  -857278969  -981967204 -1664241939
#> [139] -1316741942  -116699197  1356777000  -660818423   549796406  2082102079
#> [145]   318522484  1206957285 -1982497694 -1280863109  -197712512 -2104457343
#> [151]  1813974862   938751863  1826125132 -1667839523  -690168070   623591475
#> [157]   907542488 -1995581447  1275500390  -832738641  1776832804  -330576427
#> [163]  -537090926  -700632341  -578636496 -1173452943  1106522750  1771754727
#> [169]  1119041532 -1151885107   421735722   396381347  1541711240  -670534167
#> [175]  -492027754 -2083695073  1324124628 -1069669691  1011538114  -307008165
#> [181]  -493408032  1366131553  1203708334  1248972375   854828716  -750832707
#> [187] -1829107878  1810986259   677361464  -399798311 -1324796474  1910578575
#> [193] -1118572924  -841567307  1875056882 -2140060725  -554723184  1227651921
#> [199]  1446173918  2026906567  1950663004  1513915053  -674992758   795195779
#> [205]  -664521496 -1814682167  1748006646  1540342015  -752255180 -1293647707
#> [211]  1546151202  1050523195  -590943168  -779456703  1120035854 -1194535625
#> [217]   940673036  1375636893   598040506  1359223283   797777560  1562851257
#> [223]  -839580634  1653620847 -2127021084 -1862891115   603831634  1914684587
#> [229]  -988271632   981884721   262190910  1681842855  1518664380   938759309
#> [235] -1954554390   294884963   674591816  1576912297  -320217770  1960418271
#> [241]   990586004   -44315003  1518739842  1855876891   262035360  -471905503
#> [247]   465622638  -625128425   311040364  -179513475   754378778  1939549907
#> [253] -1552402952   939052953  1147273862 -1088236721 -1494402748  -249344139
#> [259]   868520370    25212299  1926524752   756298513  1389740446  -241233529
#> [265] -1580473316  -822667667  1584234058 -1419648189   422601640    74929545
#> [271]  -126848074   463660735  1307146740 -1189344155 -1266938398  -504522757
#> [277] -1812630784  1701058305  1565685966  1677406967    78994124  1436684637
#> [283]  -553213830 -1896959053  1207500120  1150834553    31994086 -2108631505
#> [289]  1271588516  -439024299  -500197870   550245995 -1224973648 -1144129807
#> [295]  -798360578  1118351463 -1569620612  1438435405   111496874   101228579
#> [301]  -450034936  -784673431  1583102486  2028183967   -14909612  1000159813
#> [307]  -215864766 -1732038437  1856257632   789631713  1578060590  1681820119
#> [313]   -51688404  -948552899  -169047846  2050402451  1290237112  -889335975
#> [319]  1075810118 -2100147953 -1188477948 -1568429259  2119617138  1580853067
#> [325]  1281885712  2108402385    83191390  -720126137  1624098524 -1094882771
#> [331]  -968929526  1186974979   839078504 -1975499447  1044721782 -1856779137
#> [337]  1845245108   374822949 -1416595806   794245563 -1870480960   366837441
#> [343]  1083133326  1195331767 -1786516084  1531008285 -1178558150   682298739
#> [349]  1310432280 -1893648583 -2069881434  1935617007  1684033892 -1767423723
#> [355]  1666330322  -469620693  -638625424    94719665   957350078 -2103951833
#> [361] -1925660612 -1200554995  1800634218 -1363187229   294343112  1904190761
#> [367]    63133398  1168861023  -438265324   381838853  2128540418  -572411237
#> [373]  -697768672  -356377951  -211124242  -726300777   349650668  -579117315
#> [379]   -23402086 -1450974637  1399681912  -788885735 -1593713658  -491815217
#> [385]  -288878908  1840760565  -130432206  2019350795  -227910448  -491593071
#> [391] -2125346018 -2131874553 -2079692388 -1387299347  1341775818 -1690339645
#> [397]    27066664  1156642057  1718529334  1386377791  -568217740  1180066789
#> [403]   438673250  2023398267   154058880  2058790529   825811534   811150967
#> [409]  1832730700  -694396707   613638650   670639923  -793445672  1283577593
#> [415] -1294153114   997933487   632847396   254622933 -1339717742  2105670123
#> [421]   347148336 -1613994383 -1101871746  1640860647  1301988092 -1009717299
#> [427]  1514827818 -2055736413  -334471032  1049375977  1776235414  1357966623
#> [433]   -99126060  -359968315   914127810  1874457691  -575910944 -1900862879
#> [439] -2137885522  -639482537  1024324012 -1961082179   398592602  -347939821
#> [445] -1513475528   967773913   600369350  -998607729    42572164 -1635802443
#> [451]   170753010  -240547125 -1415875696 -1008084399 -1666519074   317610695
#> [457] -1639855012  -563261011   -61001590    44097667   649949160   340354249
#> [463]  1571613174 -1253124097   152693300 -2066141275 -1728346078  -814236357
#> [469]  -189167808  -340816319   873412366 -1391932361  -828288244   -76342115
#> [475]  1346298554  1452867827   512039320  1283077817 -1453443290 -1603987601
#> [481] -1433180444  2066151573 -1855348654  1863027627   234981104  -771539407
#> [487] -1796060610  -569861721  -698907204 -1684233331   877273322 -1007534749
#> [493]  1937518408  -132086615  -579874730  -797691169   109121428  -759693947
#> [499]   314229890  1074525723  -617712992  1386474017  1783048558  -511393001
#> [505]   407856236  -468130179  -769529062  -382495277  -222449416 -1260695911
#> [511]  1161082246  -839701937  1865278532  1083915893  -588122958   736099467
#> [517]  2126203472  1225822737  -339684194  1758742663    36957980  1440146797
#> [523] -1943453366 -1767634365     2399912 -1744202615  -792729930  -820445761
#> [529]   430236916  -845169819 -2133708574    95330043   165875200 -2138556927
#> [535]   231885774   185477623 -1133500980 -1115316131   763569018  1045076659
#> [541]  1179383896   516576889  1155818470   668773679  -844033628  -967543723
#> [547] -1881245422   -94446229   748731824 -1542859279 -1313960194 -1257674905
#> [553]  -634451844   596907853   437424554  1676560163  1720630792   503092329
#> [559]  1798647062 -1175111521 -1780651436 -1425512123  -866529982     9943003
#> [565]  -441493152   777286113  -716661202   425525463   156997420 -1137620419
#> [571]  2121960410   219549587 -1459097672 -1304474023  1107640902  1691983887
#> [577]  1869934340   533371445  1497836914  1220554315   847896848  1508313553
#> [583] -1017939618   543159879 -1029647908  -682870483 -2145512950   822670339
#> [589] -1399681688   804357193   744989558  1975575423   -92102732  -597031131
#> [595]  -362178142 -1392557893 -1083484992   285253057  1128407182  1479100343
#> [601]  -110511988  -795614179  1830822970   684586099   382310168   322057785
#> [607]   573526182   396493551   701594724 -1670008811  -196865582   557575979
#> [613] -1756449680  -576705103  -918056002  1687156007  -874427588    -8959219
#> [619]  -329006486   522981603  1141378248  -270506967  -537966122 -1020002721
#> [625]  -219380460   255628549

toy_df_4col <- codes_to_counts(BradleyTerryScalable::toy_data, c("W1", "W2", "D"))
toy_btdata <- btdata(toy_df_4col)
fit2a <- btfit(toy_btdata, 1, subset = function(x) "Amy" %in% names(x))

#> Error in select_components(btdata, subset): The subset condition has removed all components

simulate(fit2a, result_class = "btdata")

#> Error in simulate.btfit(fit2a, result_class = "btdata"): simulate.btfit cannot be used for multi-component data/models