Fitted Method for "btfit"

fitted.btfit returns the fitted values from a fitted btfit model object.

# S3 method for btfit
fitted(object, subset = NULL, as_df = FALSE, ...)

Arguments

object
An object of class "btfit", typically the result ob of ob <- btfit(..). See btfit.

subset
A condition for selecting one or more subsets of the components. This can either be a character vector of names of the components (i.e. a subset of names(object$pi)), a single predicate function (that takes a vector of object$pi as its argument), or a logical vector of the same length as the number of components, (i.e. length(object$pi)).

as_df
Logical scalar, determining class of output. If TRUE, the function returns a data frame. If FALSE (the default), the function returns a matrix (or list of matrices). Note that setting as_df = TRUE can have a significant computational cost when any of the components have a large number of items.

...
Other arguments

Value

If as_df = FALSE and the model has been fit on the full dataset, returns a matrix where the \(i,j\)-th element is the Bradley-Terry expected value \(m_{ij}\) (See Details). Otherwise, a list of such matrices is returned, one for each fully-connected component. If as_df = TRUE, returns a five-column data frame, where the first column is the component that the two items are in, the second column is item1, the third column is item2, the fourth column, fit1, is the expected number of times that item 1 beats item 2 and the fifth column, fit2, is the expected number of times that item 2 beats item 1. If btdata$wins has named dimnames, these will be the colnames for columns one and two. Otherwise these colnames will be item1 and item2. See Details.

Details

Consider a set of \(K\) items. Let the items be nodes in a graph and let there be a directed edge \((i, j)\) when \(i\) has won against \(j\) at least once. We call this the comparison graph of the data, and denote it by \(G_W\). Assuming that \(G_W\) is fully connected, the Bradley-Terry model states that the probability that item \(i\) beats item \(j\) is $$p_{ij} = \frac{\pi_i}{\pi_i + \pi_j},$$ where \(\pi_i\) and \(\pi_j\) are positive-valued parameters representing the skills of items \(i\) and \(j\), for \(1 \le i, j, \le K\).

The expected, or fitted, values under the Bradley-Terry model are therefore: $$m_{ij} = n_{ij}p_{ij},$$

where \(n_{ij}\) is the number of comparisons between item \(i\) and item \(j\).

The function btfit is used to fit the Bradley-Terry model. It produces a "btfit" object that can then be passed to fitted.btfit to obtain the fitted values \(m_{ij}\). Note that the Bradley-Terry probabilities \(p_{ij}\) can be calculated using btprob.

If \(G_W\) is not fully connected, then a penalised strength parameter can be obtained using the method of Caron and Doucet (2012) (see btfit, with a > 1), which allows for a Bradley-Terry probability of any of the \(K\) items beating any of the others. Alternatively, the MLE can be found for each fully-connected component of \(G_W\) (see btfit, with a = 1), and the probability of each item in each component beating any other item in that component can be found.

References

Bradley, R. A. and Terry, M. E. (1952). Rank analysis of incomplete block designs: 1. The method of paired comparisons. Biometrika, 39(3/4), 324-345.

Caron, F. and Doucet, A. (2012). Efficient Bayesian Inference for Generalized Bradley-Terry Models. Journal of Computational and Graphical Statistics, 21(1), 174-196.

See also

btfit, btprob

Examples

citations_btdata <- btdata(BradleyTerryScalable::citations)
fit1 <- btfit(citations_btdata, 1)
fitted(fit1)

#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>               citing
#> cited             JRSS-B Biometrika       JASA Comm Statist
#>   JRSS-B       188.00000   286.4629  317.26219     281.7444
#>   Biometrika   218.53715   714.0000  505.35443     725.0322
#>   JASA         149.73781   312.6456 1072.00000     812.2472
#>   Comm Statist  11.25564    37.9678   68.75276     425.0000

fitted(fit1, as_df = TRUE)

#> # A tibble: 6 x 5
#>      component      cited       citing     fit1      fit2
#>          <chr>      <chr>        <chr>    <dbl>     <dbl>
#> 1 full_dataset     JRSS-B   Biometrika 286.4629 218.53715
#> 2 full_dataset     JRSS-B         JASA 317.2622 149.73781
#> 3 full_dataset Biometrika         JASA 505.3544 312.64557
#> 4 full_dataset     JRSS-B Comm Statist 281.7444  11.25564
#> 5 full_dataset Biometrika Comm Statist 725.0322  37.96780
#> 6 full_dataset       JASA Comm Statist 812.2472  68.75276

toy_df_4col <- codes_to_counts(BradleyTerryScalable::toy_data, c("W1", "W2", "D"))
toy_btdata <- btdata(toy_df_4col)
fit2a <- btfit(toy_btdata, 1)
fitted(fit2a)

#> $`2`
#> 3 x 3 sparse Matrix of class "dgCMatrix"
#>        player2
#> player1       Han       Gal       Fin
#>     Han .         1.1406148 0.8586132
#>     Gal 0.8593852 .         1.6412871
#>     Fin 0.1413868 0.3587129 .        
#> 
#> $`3`
#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>        player2
#> player1       Cyd       Amy       Ben       Dan
#>     Cyd .         1.2728582 .         0.7259617
#>     Amy 0.7271418 .         0.5684605 1.2042516
#>     Ben .         0.4315395 .         1.0692677
#>     Dan 0.2740383 0.7957484 0.9307323 .        
#> 

fitted(fit2a, as_df = TRUE)

#> # A tibble: 8 x 5
#>   component player1 player2      fit1      fit2
#>       <chr>   <chr>   <chr>     <dbl>     <dbl>
#> 1         2     Han     Gal 1.1406148 0.8593852
#> 2         2     Han     Fin 0.8586132 0.1413868
#> 3         2     Gal     Fin 1.6412871 0.3587129
#> 4         3     Cyd     Amy 1.2728582 0.7271418
#> 5         3     Amy     Ben 0.5684605 0.4315395
#> 6         3     Cyd     Dan 0.7259617 0.2740383
#> 7         3     Amy     Dan 1.2042516 0.7957484
#> 8         3     Ben     Dan 1.0692677 0.9307323

fitted(fit2a, subset = function(x) "Amy" %in% names(x))

#> $`3`
#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>        player2
#> player1       Cyd       Amy       Ben       Dan
#>     Cyd .         1.2728582 .         0.7259617
#>     Amy 0.7271418 .         0.5684605 1.2042516
#>     Ben .         0.4315395 .         1.0692677
#>     Dan 0.2740383 0.7957484 0.9307323 .        
#> 

fit2b <- btfit(toy_btdata, 1.1)
fitted(fit2b, as_df = TRUE)

#> # A tibble: 12 x 5
#>       component player1 player2      fit1       fit2
#>           <chr>   <chr>   <chr>     <dbl>      <dbl>
#>  1 full_dataset     Cyd     Amy 1.2677729 0.73222715
#>  2 full_dataset     Eve     Gal 0.8811003 0.11889971
#>  3 full_dataset     Han     Gal 1.1718336 0.82816645
#>  4 full_dataset     Eve     Ben 0.9108535 0.08914648
#>  5 full_dataset     Amy     Ben 0.5857476 0.41425236
#>  6 full_dataset     Eve     Dan 0.9196439 0.08035613
#>  7 full_dataset     Cyd     Dan 0.7327768 0.26722321
#>  8 full_dataset     Amy     Dan 1.2259480 0.77405202
#>  9 full_dataset     Ben     Dan 1.0566477 0.94335225
#> 10 full_dataset     Eve     Fin 0.9671479 0.03285215
#> 11 full_dataset     Han     Fin 0.8489711 0.15102890
#> 12 full_dataset     Gal     Fin 1.5978031 0.40219689