Simulate data from a specified Bradley-Terry model

Simulate data from a specified Bradley-Terry model

simulate_BT(pi, N, nsim = 1, seed = NULL, result_class = c("sparseMatrix",
  "btdata"))

# S3 method for btfit
simulate(object, nsim = 1, seed = NULL,
  result_class = c("sparseMatrix", "btdata"), ...)

Arguments

pi
a numeric vector, with all values finite and positive.

N
a symmetric, numeric matrix with dimensions the same as length(pi). The elements should be numeric representations of non-negative integers.

nsim
a scalar integer, the number of datasets to be generated.

seed
an object specifying if and how the random number generator should be initialized (‘seeded’). For details see simulate.

result_class
a character vector specifying whether the generated datasets should be of class "sparseMatrix" or of class "btdata". If not specified, the first match among those alternatives is used.

object
An object of class "btfit", typically the result of ob of ob <- btfit(..). This object must only have one component, i.e. length(object$pi) == 1.

...
Other arguments

Value

a list of length nsim of simulated datasets. If result_class = "sparseMatrix", the datasets are sparse matrices with the same dimensions as N. If result_class = "btdata" then the datasets are "btdata" objects. See btdata

See also

btfit, btdata

Examples

citations_btdata <- btdata(BradleyTerryScalable::citations)
fit1 <- btfit(citations_btdata, 1)
simulate(fit1, nsim = 2)

#> $sim_1
#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>               citing
#> cited          JRSS-B Biometrika JASA Comm Statist
#>   JRSS-B            .        302  301          289
#>   Biometrika      203          .  500          725
#>   JASA            166        318    .          809
#>   Comm Statist      4         38   72            .
#> 
#> $sim_2
#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>               citing
#> cited          JRSS-B Biometrika JASA Comm Statist
#>   JRSS-B            .        285  319          287
#>   Biometrika      220          .  516          724
#>   JASA            148        302    .          809
#>   Comm Statist      6         39   72            .
#> 
#> attr(,"seed")
#>   [1]         403         624  -207409486 -1849856373 -1042705328  -602680303
#>   [7]   297184926   616946311  1494210844  -120371347  1128119114 -1237599165
#>  [13] -1297602392 -1017047415  2037186742  -972501057  -681963788   349461861
#>  [19]  -734926110  1606979835  1945360384   339474433   933144014  1044659191
#>  [25] -1884909620   226132573 -2045371014 -1666265933   621937240 -1679665031
#>  [31] -1422393882  -241107153 -1441743964  -997091755  1670165266 -1881845909
#>  [37]  1380190128  1552816113  1827760382  -391906969 -1738542468  -694060723
#>  [43] -1950086230  -331417309  1513572360  1425348201 -1865464042 -1012004193
#>  [49] -1819831212 -1704064187  1274447682  -527783461 -2088426656   935932897
#>  [55]   396490798   511447767  -920190676   276245565  1760200154  1920156051
#>  [61] -1036846664   322457689 -1870275514  1671885327   956930308 -1032274891
#>  [67] -1737332878  1246732363   758262544  -395555887  -362589346   270764103
#>  [73]  1118223324 -1830118611   720145418  -292379133   601888616   856360521
#>  [79]  1970191734  1724036479  -492713548  2088806693  -456961118  1867199163
#>  [85]   595992256  1622564801   346585742 -1817093705 -1605754228  1101710877
#>  [91]   132980282 -2119948685  1289330968   948713529 -1621300570  1073239279
#>  [97]   723199588   202691093 -1922282542   191127851 -1719927184   849768369
#> [103]  1923378622 -1795389657 -1472449220   -64574195 -1899021206   510575331
#> [109] -1048930616 -1080367575   853763030 -1342255009 -1342132460 -1467730171
#> [115]  -542093310  1682058139  -791754208  2147187617 -1119212306 -2053369705
#> [121]   -77073428 -1920118787  -684333414  -109479085  1826486392  1809190937
#> [127]  1230317830   894248911 -1046649916  1826478069  1234329650 -1186229749
#> [133]  -906395184  -365657199 -1169377250  -857278969  -981967204 -1664241939
#> [139] -1316741942  -116699197  1356777000  -660818423   549796406  2082102079
#> [145]   318522484  1206957285 -1982497694 -1280863109  -197712512 -2104457343
#> [151]  1813974862   938751863  1826125132 -1667839523  -690168070   623591475
#> [157]   907542488 -1995581447  1275500390  -832738641  1776832804  -330576427
#> [163]  -537090926  -700632341  -578636496 -1173452943  1106522750  1771754727
#> [169]  1119041532 -1151885107   421735722   396381347  1541711240  -670534167
#> [175]  -492027754 -2083695073  1324124628 -1069669691  1011538114  -307008165
#> [181]  -493408032  1366131553  1203708334  1248972375   854828716  -750832707
#> [187] -1829107878  1810986259   677361464  -399798311 -1324796474  1910578575
#> [193] -1118572924  -841567307  1875056882 -2140060725  -554723184  1227651921
#> [199]  1446173918  2026906567  1950663004  1513915053  -674992758   795195779
#> [205]  -664521496 -1814682167  1748006646  1540342015  -752255180 -1293647707
#> [211]  1546151202  1050523195  -590943168  -779456703  1120035854 -1194535625
#> [217]   940673036  1375636893   598040506  1359223283   797777560  1562851257
#> [223]  -839580634  1653620847 -2127021084 -1862891115   603831634  1914684587
#> [229]  -988271632   981884721   262190910  1681842855  1518664380   938759309
#> [235] -1954554390   294884963   674591816  1576912297  -320217770  1960418271
#> [241]   990586004   -44315003  1518739842  1855876891   262035360  -471905503
#> [247]   465622638  -625128425   311040364  -179513475   754378778  1939549907
#> [253] -1552402952   939052953  1147273862 -1088236721 -1494402748  -249344139
#> [259]   868520370    25212299  1926524752   756298513  1389740446  -241233529
#> [265] -1580473316  -822667667  1584234058 -1419648189   422601640    74929545
#> [271]  -126848074   463660735  1307146740 -1189344155 -1266938398  -504522757
#> [277] -1812630784  1701058305  1565685966  1677406967    78994124  1436684637
#> [283]  -553213830 -1896959053  1207500120  1150834553    31994086 -2108631505
#> [289]  1271588516  -439024299  -500197870   550245995 -1224973648 -1144129807
#> [295]  -798360578  1118351463 -1569620612  1438435405   111496874   101228579
#> [301]  -450034936  -784673431  1583102486  2028183967   -14909612  1000159813
#> [307]  -215864766 -1732038437  1856257632   789631713  1578060590  1681820119
#> [313]   -51688404  -948552899  -169047846  2050402451  1290237112  -889335975
#> [319]  1075810118 -2100147953 -1188477948 -1568429259  2119617138  1580853067
#> [325]  1281885712  2108402385    83191390  -720126137  1624098524 -1094882771
#> [331]  -968929526  1186974979   839078504 -1975499447  1044721782 -1856779137
#> [337]  1845245108   374822949 -1416595806   794245563 -1870480960   366837441
#> [343]  1083133326  1195331767 -1786516084  1531008285 -1178558150   682298739
#> [349]  1310432280 -1893648583 -2069881434  1935617007  1684033892 -1767423723
#> [355]  1666330322  -469620693  -638625424    94719665   957350078 -2103951833
#> [361] -1925660612 -1200554995  1800634218 -1363187229   294343112  1904190761
#> [367]    63133398  1168861023  -438265324   381838853  2128540418  -572411237
#> [373]  -697768672  -356377951  -211124242  -726300777   349650668  -579117315
#> [379]   -23402086 -1450974637  1399681912  -788885735 -1593713658  -491815217
#> [385]  -288878908  1840760565  -130432206  2019350795  -227910448  -491593071
#> [391] -2125346018 -2131874553 -2079692388 -1387299347  1341775818 -1690339645
#> [397]    27066664  1156642057  1718529334  1386377791  -568217740  1180066789
#> [403]   438673250  2023398267   154058880  2058790529   825811534   811150967
#> [409]  1832730700  -694396707   613638650   670639923  -793445672  1283577593
#> [415] -1294153114   997933487   632847396   254622933 -1339717742  2105670123
#> [421]   347148336 -1613994383 -1101871746  1640860647  1301988092 -1009717299
#> [427]  1514827818 -2055736413  -334471032  1049375977  1776235414  1357966623
#> [433]   -99126060  -359968315   914127810  1874457691  -575910944 -1900862879
#> [439] -2137885522  -639482537  1024324012 -1961082179   398592602  -347939821
#> [445] -1513475528   967773913   600369350  -998607729    42572164 -1635802443
#> [451]   170753010  -240547125 -1415875696 -1008084399 -1666519074   317610695
#> [457] -1639855012  -563261011   -61001590    44097667   649949160   340354249
#> [463]  1571613174 -1253124097   152693300 -2066141275 -1728346078  -814236357
#> [469]  -189167808  -340816319   873412366 -1391932361  -828288244   -76342115
#> [475]  1346298554  1452867827   512039320  1283077817 -1453443290 -1603987601
#> [481] -1433180444  2066151573 -1855348654  1863027627   234981104  -771539407
#> [487] -1796060610  -569861721  -698907204 -1684233331   877273322 -1007534749
#> [493]  1937518408  -132086615  -579874730  -797691169   109121428  -759693947
#> [499]   314229890  1074525723  -617712992  1386474017  1783048558  -511393001
#> [505]   407856236  -468130179  -769529062  -382495277  -222449416 -1260695911
#> [511]  1161082246  -839701937  1865278532  1083915893  -588122958   736099467
#> [517]  2126203472  1225822737  -339684194  1758742663    36957980  1440146797
#> [523] -1943453366 -1767634365     2399912 -1744202615  -792729930  -820445761
#> [529]   430236916  -845169819 -2133708574    95330043   165875200 -2138556927
#> [535]   231885774   185477623 -1133500980 -1115316131   763569018  1045076659
#> [541]  1179383896   516576889  1155818470   668773679  -844033628  -967543723
#> [547] -1881245422   -94446229   748731824 -1542859279 -1313960194 -1257674905
#> [553]  -634451844   596907853   437424554  1676560163  1720630792   503092329
#> [559]  1798647062 -1175111521 -1780651436 -1425512123  -866529982     9943003
#> [565]  -441493152   777286113  -716661202   425525463   156997420 -1137620419
#> [571]  2121960410   219549587 -1459097672 -1304474023  1107640902  1691983887
#> [577]  1869934340   533371445  1497836914  1220554315   847896848  1508313553
#> [583] -1017939618   543159879 -1029647908  -682870483 -2145512950   822670339
#> [589] -1399681688   804357193   744989558  1975575423   -92102732  -597031131
#> [595]  -362178142 -1392557893 -1083484992   285253057  1128407182  1479100343
#> [601]  -110511988  -795614179  1830822970   684586099   382310168   322057785
#> [607]   573526182   396493551   701594724 -1670008811  -196865582   557575979
#> [613] -1756449680  -576705103  -918056002  1687156007  -874427588    -8959219
#> [619]  -329006486   522981603  1141378248  -270506967  -537966122 -1020002721
#> [625]  -219380460   255628549

toy_df_4col <- codes_to_counts(BradleyTerryScalable::toy_data, c("W1", "W2", "D"))
toy_btdata <- btdata(toy_df_4col)
fit2a <- btfit(toy_btdata, 1, subset = function(x) "Amy" %in% x)
simulate(fit2a, result_class = "btdata")

#> $sim_1
#> $wins
#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>        player2
#> player1 Cyd Amy Ben Dan
#>     Cyd   .   0   .   0
#>     Amy   2   .   1   1
#>     Ben   .   0   .   1
#>     Dan   1   1   1   .
#> 
#> $components
#> $components$`1`
#> [1] "Cyd" "Amy" "Ben" "Dan"
#> 
#> 
#> attr(,"class")
#> [1] "btdata" "list"  
#> 
#> attr(,"seed")
#>   [1]         403          24 -2089519707 -1093835428 -1319683790  1392429101
#>   [7]  -202324249   631843854  -539511320  1353072619 -1303625591   879318040
#>  [13]   774136646  1478877601 -1644826573 -1155924574 -1612494700 -1731969257
#>  [19]  -845694707   474058788 -1674022902 -1410673291   -72657713   325212118
#>  [25]  1081710464 -1118188605 -1231921823   521389008  1462334254   265391833
#>  [31]  1689205995 -1189593766  -976466148   404318751  -837617035 -1825212660
#>  [37]  1573186274      315325 -1025460169 -1853639970  2095787544  -982486661
#>  [43]   922061081  1343641000   946764118   530950609   329083267  -334890990
#>  [49]  -309281884  1177791335  1832358077  1128990772  1553173402    33392613
#>  [55]  1738486655 -1895238906   316341392 -1461736141 -1293989167 -1084706688
#>  [61]   251251038  -199237975  -361027237   976855914 -1401058804   295033423
#>  [67]  1231140613   528547772   703052434  -305520243   510577095  -706028882
#>  [73]  2077389576  1635348747    55128041 -1622058824  2080255846  1787411457
#>  [79]   -51320173   802553154 -1422732556 -1690329353  1321746157 -1152358012
#>  [85]  -764057686 -1690460907    -2386001  1579456438   736030176 -1761295069
#>  [91]  1204959809 -1338288848   517082958  1004819897    84322251  1727118458
#>  [97]  1730305404  -606472769   214872725   171765932   317487746   736686685
#> [103]  -746781353 -1892652994 -1246229384   757118811   935006905   955395848
#> [109]   584918454 -1957668111  1740918051 -1430395022  2039604676   720149895
#> [115]   651857757  1196065108 -1502727302 -1034624315  -266779617    50901926
#> [121] -1763801808 -2141575213 -1923070479  1952256800  -113943106 -1467755063
#> [127]   352987515   552218442  2003816364  -193538385  1253409893  2102928412
#> [133]   562907762  -571183891   -86682329  1801396558  1474711976  1427129771
#> [139]  -293388599  -736147752   147016582  2111946337 -1722873869  1328929378
#> [145] -1539654444  1074716119  -418916403 -1213562012   504661834  -103678283
#> [151]  2024568335   449611542   860488128  -624893565  2123659169 -1054541936
#> [157] -1105572754  2081544729  -574078293   310023834 -1294876068    80445279
#> [163]  1273172277   -82038580  1862769186 -1398549379 -1387106953  -990461282
#> [169] -1426648744 -1713829701 -1954179495   615106024   155041814   254585745
#> [175]  -705162173  -831411374   187232228   471871015  -888625539   507475828
#> [181]  -294327462  2056830373  -185566785   737161798  -378527536 -1976665357
#> [187]  -860881903  -695268800  2052148766  -660961047  -789525989 -1304724950
#> [193]  -804098740  1535520527  -319753659 -1779084292   569674322 -1049561139
#> [199]   161472903 -2058800018 -1525257144  1755321931 -1768739927  -715414792
#> [205]     7357990   423108417 -1675451181  1063069826  -750205260  1278960183
#> [211] -1847301843   129108548  1840789354 -1792728107 -1537919633  -884576138
#> [217]  1817697184  -661308317  -721073919  -601373584  1223450638  -895926151
#> [223]  -143884277  1150554426  1049858108 -1911449729  1664396501  -798977812
#> [229]  -881391038 -2069615046   366650576 -1676327828  -804498764  -162584558
#> [235]  -194488352  -875753524  -414159392 -2076660190  -687003352   866251020
#> [241]    -8341956  -256450318   358513392 -1741921468   807610968 -2139631046
#> [247] -1474853136  -713196356   138677780 -1378869054 -1280872352  1371901308
#> [253]  2142051536 -1801233374 -1719002456    73396108  -297231940  1028486194
#> [259]   550557504   -56198988 -1751399608  -411559238   854740752   220272876
#> [265]   -25693100 -1378482414   490468256   812851148  1581476704   181252290
#> [271] -2066770392   513459628 -1708974276  1235447666   100107056 -1536556508
#> [277] -1383526344  1134472410  -339861104  1590241020 -1306247276  1209819970
#> [283]  1366759936  1898349116   328939088 -1305783646   662203208  1797044748
#> [289]  1610476764 -1687362990  -212395904 -1224492204  -713322008   559237882
#> [295]  -405006512   954139692   505214324   490499730  1340488288  1229255308
#> [301] -1863537824  1818702562   975714472  1148195020 -1370516996  -983433294
#> [307]   -88287760  -597635644   364857176  1640995706   273128880  -335494724
#> [313]  1636708436 -2060594238 -1170771168   229463996   728928848   668859874
#> [319]   677036392  1111928140  -206753540  2049819378  -790233344  1091985012
#> [325]  2080673800   246677946 -1178755120  1653250540 -1379993708 -1335713070
#> [331]  1969700448 -1397119668 -1324866272  1754892930 -1454147096  1411596268
#> [337]   426985020  1913419378  1679739760  1909504292  -116747720   110963098
#> [343]   -35808816 -1397766724   922548372  -319220606  -833256000 -1311552708
#> [349]   516033680  -818918622 -1795541496  -888834804   987284892  1688419858
#> [355] -1948021888   462971668   799850792  1446872122  -451114032  1729178220
#> [361]   415539764 -1887638894  -452450976   522451276  -190371872  -278751838
#> [367]  1875002792   976422540 -1350025796 -1507819918  -120736784  -821113276
#> [373] -1402511272 -1632442310  1157165168  1417615420  1315234068   132186818
#> [379]  -687487264  1526511740  1807029968   773101986   695990312  -914470132
#> [385] -1815967300  -506158286  -144946496 -1947931596  -325402936   189527226
#> [391]   576951952   755150572  1661786708  -894532974  1348585760  -365325876
#> [397]   581550816 -1196244798 -1796191192   640489388  -183555396  -655696910
#> [403]  1798592048   264212388 -1200370248  -745220262 -1625599856  -655733892
#> [409]  -924966636 -1497773758 -1593569024 -1054188100  1969627472  -708732638
#> [415]   305797704  1754099724  1174709852  -924710446  1442002432  -632625836
#> [421]  -848778008 -1277587206  2082154192   121236268  2039492852  1689966994
#> [427]   519732576  -752985076 -1360085792 -1568886174  1680565416  1031535692
#> [433]  -176879748  -954427086 -1853267344   -39918780 -1838529576  -777170054
#> [439]  -229926864   859754428   161722964  -273403838  1743761056  1051902652
#> [445]   491391952  -991871646  1729253096   162049228  1554915836 -1300611598
#> [451]  -448240512  1801333236  1995295368   761387834     2035024 -1000870804
#> [457]  1922709043  1811168405  1145930994 -1017977536  1700997801  -625073381
#> [463]  1214436588 -1545824006   526158127 -1687098055  -854417586 -2049149540
#> [469]   476681997  1331290487   396408192  1467188190  1690758315 -1073154131
#> [475]  1731808282   117867224   317033281 -1492620301 -1728072172   719630242
#> [481]  1882755255 -1858155903  -355296634  -381134140 -1924899851  1789522527
#> [487]   178947256   -55574122   868154275  1758945125 -1244612990 -2016812048
#> [493]   147781017  1938561931 -1052658052   486796458 -1163985185  -846007191
#> [499] -1483239810 -1212930452 -1174787395   819206023   228643696  -508648946
#> [505]  -392532165 -2102471651 -1286054710  -881137688 -1954974511   871186627
#> [511]  -824727612  1678623858   191790407   151032081 -1476660074  1621117428
#> [517]   653007749  -961056049  1516318664  -731748762 -1833531245  -506939403
#> [523]  -360336942  1578023456 -2010115063   666487099 -1212138356  -599860966
#> [529]  -744464497  2051088601 -1615859858 -1430382852  1546474605 -1135218665
#> [535]  1089432032   193418686 -1509557365  1183190029  1781349434  1118555256
#> [541]   322979873 -1740291373 -2025082380   229251714  -570460393  1874197985
#> [547] -1503562970 -1510117852  -249456427 -2009941889 -2103812840   286463030
#> [553]  -965297469 -1263216891  -898299230 -1627805552   664033337  1090365675
#> [559]   692816348   -68775926 -1542123777  -813719735 -1923066658   732228236
#> [565] -1137770915  1320846439 -1535735408   -39364178 -2121450405 -2099192771
#> [571]   858550954   578752328   753795953 -1353243421   448211940   695761682
#> [577]  1507702503 -1202085711  1986908918   597439764  -782759387 -1437976785
#> [583]  -759224088   251439558 -1533859725 -1525447979 -1499602766   244180096
#> [589]  1099733097  -214174501 -1127532500  -907646918   408889071   452364665
#> [595]  -180087282  -264398884  2007531597  -647191625  2048056640 -1865171938
#> [601]  1878239979   845734509  -653176102  1157638808 -1754224383  -971870541
#> [607]  -937096748  2136844514 -1188630153  1398403265    15404102 -1738157180
#> [613]   369886517 -1811971297  1621862008  -915793066  -302456349  1414690725
#> [619]   -73656766  1091099696  1683357785  1066823115   313056828  1541374058
#> [625]  1713650335   205702693