Simulate data from a specified Bradley-Terry model

Simulate data from a specified Bradley-Terry model

simulate_BT(pi, N, nsim = 1, seed = NULL, result_class = c("sparseMatrix",
  "btdata"))

# S3 method for btfit
simulate(object, nsim = 1, seed = NULL,
  result_class = c("sparseMatrix", "btdata"), ...)

Arguments

pi
a numeric vector, with all values finite and positive.

N
a symmetric, numeric matrix with dimensions the same as length(pi). The elements should be numeric representations of non-negative integers.

nsim
a scalar integer, the number of datasets to be generated.

seed
an object specifying if and how the random number generator should be initialized (‘seeded’). For details see simulate.

result_class
a character vector specifying whether the generated datasets should be of class "sparseMatrix" or of class "btdata". If not specified, the first match among those alternatives is used.

object
An object of class "btfit", typically the result of ob of ob <- btfit(..). This object must only have one component, i.e. length(object$pi) == 1.

...
Other arguments

Value

a list of length nsim of simulated datasets. If result_class = "sparseMatrix", the datasets are sparse matrices with the same dimensions as N. If result_class = "btdata" then the datasets are "btdata" objects. See btdata

See also

btfit, btdata

Examples

set.seed(1)
n <- 6
W <- matrix(rpois(n ^ 2, lambda = 5), n, n)
N <- W + t(W)
pi <- exp(rnorm(n)/4)
simulate_BT(pi, N, seed = 6)

#> $sim_1
#> 6 x 6 sparse Matrix of class "dgCMatrix"
#>                      
#> [1,] 8  6  8  7  4  7
#> [2,] 7 12  8  7  3  5
#> [3,] 3  2 14 10  3  8
#> [4,] 5  2  4  6  7  5
#> [5,] 3  4  4  3 16  6
#> [6,] 6  3  8  1  5 12
#> 
#> attr(,"seed")
#> [1] 6
#> attr(,"seed")attr(,"kind")
#> attr(,"seed")attr(,"kind")[[1]]
#> [1] "Mersenne-Twister"
#> 
#> attr(,"seed")attr(,"kind")[[2]]
#> [1] "Inversion"
#> 

citations_btdata <- btdata(BradleyTerryScalable::citations)
fit1 <- btfit(citations_btdata, 1)
simulate(fit1, nsim = 2, seed = 1)

#> $sim_1
#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>               citing
#> cited          JRSS-B Biometrika JASA Comm Statist
#>   JRSS-B            .        286  316          285
#>   Biometrika      219          .  495          725
#>   JASA            151        323    .          817
#>   Comm Statist      8         38   64            .
#> 
#> $sim_2
#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>               citing
#> cited          JRSS-B Biometrika JASA Comm Statist
#>   JRSS-B            .        291  315          273
#>   Biometrika      214          .  511          723
#>   JASA            152        307    .          811
#>   Comm Statist     20         40   70            .
#> 
#> attr(,"seed")
#> [1] 1
#> attr(,"seed")attr(,"kind")
#> attr(,"seed")attr(,"kind")[[1]]
#> [1] "Mersenne-Twister"
#> 
#> attr(,"seed")attr(,"kind")[[2]]
#> [1] "Inversion"
#> 

toy_df_4col <- codes_to_counts(BradleyTerryScalable::toy_data, c("W1", "W2", "D"))
toy_btdata <- btdata(toy_df_4col)
fit2a <- btfit(toy_btdata, 1, subset = function(x) "Amy" %in% x)
simulate(fit2a, result_class = "btdata")

#> $sim_1
#> $wins
#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>        player2
#> player1 Cyd Amy Ben Dan
#>     Cyd   .   1   .   1
#>     Amy   1   .   1   0
#>     Ben   .   0   .   1
#>     Dan   0   2   1   .
#> 
#> $components
#> $components$`1`
#> [1] "Cyd" "Amy" "Ben" "Dan"
#> 
#> 
#> attr(,"class")
#> [1] "btdata" "list"  
#> 
#> attr(,"seed")
#>   [1]         403          48  1654269195 -1877109783  -961256264  1403523942
#>   [7]   124639233   261424787  1836448066  1034917620   -13630729   468718317
#>  [13]  1694379396  1559298986  1935866133 -1450855505  2105396150  1802260960
#>  [19]  1077391651   539731521   122505520   230898510 -1940184647  1223031755
#>  [25] -1597886342 -1854140036 -1783225921  1484611221  1365746860  -346485118
#>  [31]  1206044253  1201793367   956757054   350214264 -1324711077 -1071776071
#>  [37] -1831283960  1862871478  1826141937   268853539   607122290  1403468228
#>  [43] -1913257593  -346768547  -463801004  1625143162  1387501253  -280700897
#>  [49]   422985638   732086576  -725197869  1128169969 -1024391392 -1032957506
#>  [55]  -353082935 -1139485317   445858634  1224049580  -796824913   247650405
#>  [61]   542595100  -404965774  -423589651 -1999189721  1430762830 -1439787608
#>  [67] -1821953621  1808288457 -1668269352 -1099348090    24992353  -140278797
#>  [73]  1709299298   269218004  1210153431  -238785587  -383705244  -472446134
#>  [79]   -15672651  -538858993  1771793174   -40319552 -1632234109 -1613456479
#>  [85]  -109108336  2056208494   -22837735  -195336533   835590810    86083164
#>  [91]   517854047  -397095627 -1681163060   985204258   613283965  1720828279
#>  [97]   929066654  -293216936   348261563  1602155097  1188944360  -526197738
#> [103] -1056856175  -999184317 -1361965230 -1531694108  1775225383  1578975869
#> [109]  -305972364   499275610  -297283163 -1400910401  -628019642  -995328816
#> [115]  -416901389  -131985391   838014528   160971294   516863209 -1388986853
#> [121] -1140425174   -14278324  1980074767  1968938565  -674453508  1981972050
#> [127] -1121442867  -907997817   494782574  1629589576    57352779 -1057292375
#> [133]  2043531000 -1557378522 -2033982655 -1166473005 -1243272062  1507691188
#> [139]   519478839  -214579923   -15421884  -257117334   917217237  1604258159
#> [145] -1793932170  1404112800   895408227 -1228178175 -1129918352  1726708750
#> [151] -1514890119 -1503646709 -1385642694   530830396 -1346935937  1552658645
#> [157]   -31128340 -1631936958  1895821725   665429015  1197209982    25108280
#> [163]  1382701083 -1944495751  1482522568  1344742390  2075751985   887378019
#> [169]  2132045874   507710852 -1239616569 -1239992675  1300848148 -2078006086
#> [175]  1822724357  1779634911  -346210458    68638448  -651383277   408315825
#> [181] -2048675232  1331662846 -1134018679 -1715119941 -1026157430  -262084756
#> [187] -1837320977  -648290523  -685012516  -699648078 -1959870803  -939002265
#> [193]  1368704398  1604498024  -270455189   878774281 -1496090984  1651241414
#> [199]  -719838943  -170397261  1455721762   967288596  1974706071 -1131741299
#> [205]  1944879780   693669002   967713525   849937743  1265996374   507097344
#> [211]  -355466173  2106394593   467134288  1892604334   231167833   250371435
#> [217]  1139934426  -641921636  -858574433  -474934283 -2013707636 -1380890782
#> [223]   119355197   993660855 -2002084770 -1013984872  -379702021   925663385
#> [229]   -31002328  -442860704 -1075363102   641540264  1223098572 -1416033796
#> [235]  1656623026  -797919760  1826465220   503635800   479382394  -180521552
#> [241]   335149500  -762107820 -1027404862 -1827060960  1940004796   -55285168
#> [247]   820174306 -1241009816   780750156  1441990908   182952690   711513856
#> [253] -1637546380  2029275144  1579137466 -1032808496  -740935188  -653815916
#> [259]   804057810 -1415082400  1411310924     1021216  -350965118  1885104616
#> [265]   987875308  -131369412  1998212722 -1838333072   848054564  1637511736
#> [271] -1985453670   185338320  1063452092 -1501837164  -244441982  -655956544
#> [277]  1724949308  2018636944   664617762   824380936  -454972148  1579726236
#> [283] -1666261486  -245134464   802374420  1526755624   260631610 -1205064752
#> [289]  -361627028   937556532  -494404974   -99312288 -1390302900  1614862304
#> [295]   774509986  1360092584  -193173876 -1314447940 -1663111566 -1553441808
#> [301]  1474524740 -1146038184 -2113923014  1375914096 -1694107588  1286052116
#> [307]  -244301118 -1394274592 -1799518084  1647537872  -826656350   363032616
#> [313]   887266060  1523945916 -1606634702 -1705936192  1284490292  -611903800
#> [319]   -82717510  1223915152  1740148460 -1417786796   261071506  -232439520
#> [325]  -735589940 -1901903136    53278914 -1123929048  -739412052  1446503100
#> [331] -1800705038  1762987568 -1902032988 -1903022664   152627034   717533328
#> [337]   246260604  1132269844 -1594365630   343599360 -1585484356   763185488
#> [343] -1647048414  -696819128   134752268 -1533188516 -1147008558  1720888832
#> [349] -1992442540 -1205072664  1506019578   603356880  1553058604   599599348
#> [355]  -568221294   -26303136  1900303372  -891679520 -1327582622  -770317144
#> [361] -2141666228  1110458236 -1984349902   163365488  1002681156 -1830044712
#> [367]  -482759814  1341773872 -1905328196 -2080945580   830894146  1654662816
#> [373]   550937276  -155556912  -506636958  1307590376  1514790092  1301985788
#> [379] -2135581198  1363303552  2121865716  1892557448  1059986234  1182360912
#> [385] -1907645332 -1564824172  1223534674 -1003414048  1916359884  -428286304
#> [391]  -449023614 -1081851672  1379461868   -76178116    39708786 -1623293328
#> [397]   506321444  2118223928 -1103890150  -418716464 -1680746180 -1496452204
#> [403]   625145730  1623387072  1943565884  1450672528  1732028450  1481551112
#> [409]   487458956   872206108  -748118894  2140900736  -935907308   548710184
#> [415] -1376221638  1162922448 -1448468116   593599412 -1498189294 -1658751008
#> [421]   462232012   124672224   557945890  1357773608   781625868  1816482620
#> [427]  -142301198  -510500112  1926556228   -13969832  -233818310 -1136707600
#> [433]   460294844  -416275180  1083468994  -781480864  1355603324  1851695056
#> [439]   578619682  -798888280  1535516556  -172522308  1128772402  -958671296
#> [445] -1018973004  1397935176 -1571451462  1620720144  -804320020  -120377004
#> [451]   539237138  1130743200  -645778740  1151912544  1355687874  1031070504
#> [457] -1306313446   142889871   589573337 -1228012690 -1810984196 -1041194899
#> [463]  -926225385 -1469796896  1488304574 -1061851253  -338333683 -2027587526
#> [469]  -665470856  1572082721  -864508205  -357575180   -92260734  -727936233
#> [475] -1218739743 -1034735834  1959295524 -2018968363   970828927  2088100120
#> [481]  1582403638 -1341110589 -1107547387   214011554  -657752944 -1164080071
#> [487]  -795818773  -913533476   200285194   752745215    63615817 -1335801634
#> [493] -1217612148  -805518755 -1818050457 -1804608112   621505966   918960219
#> [499]   676442173  -402544982 -1789574328   224036721  -764759837  -670177308
#> [505]   261209874  -893263129  1281983665   362571510  1179824404  1440671269
#> [511]  -385713361  -963744536  2009648582 -1648018317  -719300907  1092938930
#> [517]   163127424   830016617  -184674085   343641132   154989626  -583638289
#> [523]   424254841   484102670 -1823956004   550505549  -397533257  -309821120
#> [529] -1833768930   947035883  1101549165  1819580122  1254016664   600796417
#> [535]  -119647565   507059668  1219230946   981013879   178359489 -1798344634
#> [541]   519307140  1302534453   -26966241  -277288328 -1009389738   145067491
#> [547]   844768165 -1768828350 -1545351120  -672619431 -1158866485    15817276
#> [553]  -881306518   -50051425 -1092056279  1196083134   248280748   296336381
#> [559]   891261767   207606832 -1985825458 -2102088325  1662966493 -1304487158
#> [565]   873339432   806441617  1094718083  1123275652   840202674   725952263
#> [571]  -597236399 -1409356586  -114177612  1521819077   401542415  1975999368
#> [577]   -10714458 -1167764909  1721318325 -2086876910   742005472   969609033
#> [583]  -890084997 -1277850548  -971870758  1665781711   -20106855  1492599214
#> [589]  1206353596   231206701  1215581399  -910904032  -626793858   303180875
#> [595]   704691789  1846420602 -1294941000 -2057982367   787944467 -1144544716
#> [601]  -630963134  1707080791   636622753  -866277274   434270308   154340245
#> [607]   835108031 -1921487784   511924342 -1737570173   907225285   803644898
#> [613]  -625275824 -1045666439 -1073554773 -1362001380  -645341878   459588927
#> [619] -1339341943   -24970338 -2003084724  -426558435 -1332570713  -956733232
#> [625]   355793518  1450597184